Moved debate database storage tests to new file
